refactor: localize worker runtime contracts
