+++ deploy/api-worker-shell/types.ts
@@ -167,87 +167,3 @@ export interface SupabaseCacheState<T> {
   pending: Promise<T> | null;
   value: T | null;
 }
-
-export interface WorkerReviewReadService {
-  handleReviewFeed(request: Request, env: WorkerEnv, url: URL): Promise<Response>;
-  handleReviews(request: Request, env: WorkerEnv, url: URL): Promise<Response>;
-  handleReviewDetail(request: Request, env: WorkerEnv, reviewId: string): Promise<Response>;
-  loadSingleReview(env: WorkerEnv, reviewId: string, sessionUserId?: string | null): Promise<WorkerReview | null>;
-  mapReviewRows(
-    feedRows: WorkerJsonRecord[],
-    commentRows: WorkerJsonRecord[],
-    likeRows: WorkerJsonRecord[],
-    usersById: Map<string, WorkerJsonRecord>,
-    placesByPositionId: Map<string, WorkerPlace>,
-    stampRowsById: Map<string, WorkerJsonRecord>,
-    routeRows: WorkerJsonRecord[],
-    likedFeedIds: Set<string>,
-  ): WorkerReview[];
-}
-
-export interface WorkerCommunityRouteService {
-  handleCommunityRoutes(request: Request, env: WorkerEnv, url: URL): Promise<Response>;
-  handleCreateUserRoute(request: Request, env: WorkerEnv): Promise<Response>;
-  handleMyRoutes(request: Request, env: WorkerEnv): Promise<Response>;
-  handleToggleCommunityRouteLike(request: Request, env: WorkerEnv, routeId: string): Promise<Response>;
-  loadCommunityRoutes(env: WorkerEnv, options?: WorkerJsonRecord): Promise<WorkerJsonRecord[]>;
-}
-
-export interface WorkerMyService {
-  handleMyComments(request: Request, env: WorkerEnv, url: URL): Promise<Response>;
-  handleMySummary(request: Request, env: WorkerEnv): Promise<Response>;
-}
-
-export interface WorkerAdminService {
-  handleAdminSummary(request: Request, env: WorkerEnv): Promise<Response>;
-  handleAdminImportPublicData(request: Request, env: WorkerEnv): Promise<Response>;
-  handleAdminPlaceVisibility(request: Request, env: WorkerEnv, placeId: string): Promise<Response>;
-}
-
-export interface WorkerStampService {
-  handleToggleStamp(request: Request, env: WorkerEnv): Promise<Response>;
-}
-
-export interface WorkerNotificationInsertResult extends WorkerJsonRecord {
-  notification_id?: string | number | null;
-}
-
-export interface WorkerNotificationCreatePayload extends WorkerJsonRecord {
-  userId: string;
-  type: string;
-  title: string;
-  actorUserId?: string | null;
-  body?: string;
-  reviewId?: string | number | null;
-  commentId?: string | number | null;
-  routeId?: string | number | null;
-  metadata?: WorkerJsonRecord;
-}
-
-export interface WorkerReviewInteractionDeps {
-  badgeByMood: Record<string, string>;
-  countUnreadNotifications(env: WorkerEnv, userId: string): Promise<number>;
-  createUserNotification(env: WorkerEnv, payload: WorkerNotificationCreatePayload): Promise<WorkerNotificationInsertResult | null>;
-  loadBaseData(env: WorkerEnv, sessionUserId?: string | null): Promise<WorkerBaseData>;
-  loadNotificationById(env: WorkerEnv, notificationId: string | number): Promise<WorkerJsonRecord | null>;
-  loadSingleReview(env: WorkerEnv, reviewId: string, sessionUserId?: string | null): Promise<WorkerReview | null>;
-  publishNotificationEvent(env: WorkerEnv, userId: string, event: string, payload: WorkerJsonRecord): Promise<void>;
-  readSessionUser(request: Request, env: WorkerEnv): Promise<WorkerSessionUser | null>;
-}
-
-export interface RouteRuntime {
-  adminService: WorkerAdminService;
-  buildReviewInteractionDeps: () => WorkerReviewInteractionDeps;
-  communityRouteService: WorkerCommunityRouteService;
-  loadBaseData: (env: WorkerEnv, sessionUserId?: string | null) => Promise<WorkerBaseData>;
-  loadStaticBaseRows: (env: WorkerEnv) => Promise<WorkerStaticBaseRows>;
-  mapCourses: (
-    courseRows: SupabaseCourseRow[],
-    coursePlaceRows: SupabaseCoursePlaceRow[],
-    placesByPositionId: Map<string, WorkerPlace>,
-  ) => WorkerCourse[];
-  mapPlace: (row: SupabaseMapRow) => WorkerPlace;
-  myService: WorkerMyService;
-  reviewReadService: WorkerReviewReadService;
-  stampService: WorkerStampService;
-}
